+++ HrApp.Domain/Entities/Paid.cs
@@ -5,5 +5,5 @@ public class Paid
     public Guid Id { get; set; }
     public Guid UserId { get; set; }
     public virtual User User { get; set; } = default!;
-    public float BaseSalary { get; set; }
+    public double BaseSalary { get; set; }
 }
